WIP Build 2
-Stressed tested many data model functionalities and added extra
features
-added a crude test to processing transactions by adding a listener to
the Buy button of every card.  I would imagine that this interface would
obviously be more complex than that, but it is simply for testing
-my concept for a real ordering system is maybe have a fragment appear
when you press the menu button that contains a list of all menu items in
a listview which has checkboxes next to each of them. once the buy
button is pressed, it adds all checked items into one transaction and
commits it to the datamanager.