Filter ALMOST DONE -- PLEASE read!
Note: Two things that are broken in this build
-somehow I managed to break the menu. The menu doesnt populate any menu
entries any more. When it was still working, it was working wonderfully.
-another thing I would like to implement is to redraw the UserActivity
at any time when the menu is exited.

(ALMOST) completed features:
-the filter appears to work in real time! if I log transactions through
the interface, it will sort the results in real time, given that I
refresh the screen with a re-log.
-most of the app crashing issues were due to the removal of the local
restaurant array in the onPause.  Im wondering if this has somethin to
do with the fact the menu doesnt populate entries?