Filter complete, only a handful of TODOs for demotime!
Filter fully works within the user experience and displays results based
on comparisons between each restaurant's menu characteristics and the
user's cached preference data. Also fixed a problem with merge not
copying the menu correctly.

TODOs:
-allow the buy button on the menu fragment to refresh the UserActivity
to reflect the changes in the cache state
-put a website button on each card, I wrote the hooks for the Restaurant
database to store the URL as a string in each Restaurant instance, and
added it as a parameter in the JSON constructor. Can be accessed with
Restaurant.getURL();(added a TODO comment)
-I would like to get a Toast to display on the screen when a transaction
is logged that simply displays a thank you message and a total
transaction cost (added a TODO comment)
-On that note, having the menu hold the individual item's values would
be good too.